Move primitive types to 'astar-primitives'.
