+++ pallets/ethereum-checked/Cargo.toml
@@ -11,7 +11,6 @@ repository.workspace = true
 parity-scale-codec = { workspace = true }
 scale-info = { workspace = true }
 
-ethereum = { workspace = true }
 ethereum-types = { workspace = true }
 fp-ethereum = { workspace = true }
 fp-evm = { workspace = true }
@@ -23,8 +22,11 @@ frame-system = { workspace = true }
 sp-runtime = { workspace = true }
 sp-std = { workspace = true }
 
+astar-primitives = { workspace = true }
+
 [dev-dependencies]
 hex = { workspace = true }
+ethereum = { workspace = true }
 pallet-balances = { workspace = true }
 pallet-ethereum = { workspace = true }
 pallet-evm = { workspace = true }
@@ -36,8 +38,8 @@ default = ["std"]
 std = [
 	"parity-scale-codec/std",
 	"scale-info/std",
+    "ethereum/std",
 	"ethereum-types/std",
-	"ethereum/std",
 	"fp-ethereum/std",
 	"fp-evm/std",
 	"pallet-evm/std",
@@ -50,6 +52,7 @@ std = [
 	"pallet-timestamp/std",
 	"pallet-evm/std",
 	"pallet-ethereum/std",
+    "astar-primitives/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
