+++ primitives/Cargo.toml
@@ -9,6 +9,12 @@ repository.workspace = true
 [dependencies]
 # third-party dependencies
 log = { workspace = true }
+parity-scale-codec = { workspace = true }
+scale-info = { workspace = true }
+
+# Ethereum/Frontier dependencies
+ethereum = { workspace = true, features = [ "with-codec" ] }
+ethereum-types = { workspace = true }
 
 # Substrate dependencies
 frame-support = { workspace = true }
@@ -28,6 +34,10 @@ pallet-xc-asset-config = { workspace = true }
 default = ["std"]
 std = [
 	"log/std",
+    "parity-scale-codec/std",
+    "scale-info/std",
+    "ethereum-types/std",
+	"ethereum/std",
 	"frame-support/std",
 	"sp-std/std",
 	"sp-runtime/std",
