+++ primitives/src/ethereum_checked.rs
@@ -1,0 +1,89 @@
+// This file is part of Astar.
+
+// Copyright (C) 2019-2023 Stake Technologies Pte.Ltd.
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+// Astar is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Astar is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Astar. If not, see <http://www.gnu.org/licenses/>.
+
+use parity_scale_codec::{Decode, Encode};
+use scale_info::TypeInfo;
+
+use ethereum::{
+    AccessListItem, EIP1559Transaction, TransactionAction, TransactionV2 as Transaction,
+};
+use ethereum_types::{H160, H256, U256};
+use frame_support::{dispatch::DispatchResultWithPostInfo, pallet_prelude::*};
+use sp_std::prelude::*;
+
+/// The checked Ethereum transaction.
+#[derive(Encode, Decode, Clone, PartialEq, Eq, RuntimeDebug, TypeInfo)]
+pub struct CheckedEthereumTx {
+    /// Gas limit.
+    pub gas_limit: U256,
+    /// Action type, either `Call` or `Create`.
+    pub action: TransactionAction,
+    /// Amount to transfer.
+    pub value: U256,
+    /// Input of a contract call.
+    pub input: Vec<u8>,
+    /// Optional access list, specified in EIP-2930.
+    pub maybe_access_list: Option<Vec<(H160, Vec<H256>)>>,
+}
+
+impl CheckedEthereumTx {
+    pub fn into_ethereum_tx(&self, nonce: U256, chain_id: u64) -> Transaction {
+        let access_list = if let Some(ref list) = self.maybe_access_list {
+            list.iter()
+                .map(|(address, storage_keys)| AccessListItem {
+                    address: *address,
+                    storage_keys: storage_keys.clone(),
+                })
+                .collect()
+        } else {
+            Vec::new()
+        };
+
+        Transaction::EIP1559(EIP1559Transaction {
+            chain_id,
+            nonce,
+            max_fee_per_gas: U256::zero(),
+            max_priority_fee_per_gas: U256::zero(),
+            gas_limit: self.gas_limit,
+            value: self.value,
+            action: self.action,
+            input: self.input.clone(),
+            access_list,
+            odd_y_parity: true,
+            r: dummy_rs(),
+            s: dummy_rs(),
+        })
+    }
+}
+
+/// Dummy signature for all transactions.
+fn dummy_rs() -> H256 {
+    H256::from_low_u64_be(1u64)
+}
+
+/// Transact an checked Ethereum transaction. Similar to `pallet_ethereum::Transact` but
+/// doesn't require tx signature.
+pub trait CheckedEthereumTransact {
+    /// Transact an checked Ethereum transaction in XVM.
+    fn xvm_transact(source: H160, checked_tx: CheckedEthereumTx) -> DispatchResultWithPostInfo;
+}
+
+/// Mapping from `Account` to `H160`.
+pub trait AccountMapping<AccountId> {
+    fn into_h160(account: AccountId) -> H160;
+}
